Add Loading Indicator
Added a large loading indicator when you press 'generate' in the kitchen. While demoing our project, we noticed a large number of volunteers would click 'generate' many times because they did not know if it worked or not, when in reality it was working---they simply had to wait a few more seconds for the generation to finish. Without a clear 'loading' indicator, users did not know if it was working or not. The new loading screen fixes this.